requirements.txt
amqp==2.2.2
appnope==0.1.0
beautifulsoup4==4.6.0
billiard==3.5.0.3
celery==4.1.0
certifi==2017.7.27.1
chardet==3.0.4
decorator==4.1.2
Django==1.11.6
django-constance==2.0.0
django-picklefield==1.0.0
gunicorn==19.7.1
idna==2.6
ipython==6.2.1
ipython-genutils==0.2.0
jedi==0.11.0
kombu==4.1.0
meld3==1.0.2
parso==0.1.0
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.15
ptyprocess==0.5.2
Pygments==2.2.0
pytz==2017.2
redis==2.10.6
requests==2.18.4
simplegeneric==0.8.1
six==1.11.0
supervisor==4.0.0.dev0
traitlets==4.3.2
urllib3==1.22
vine==1.1.4
wcwidth==0.1.7
